move quicklinks
