+++ src/features/quickAccess/AddQuickLinkForm.tsx
@@ -1,6 +1,7 @@
 import { z } from "zod"
-import NormalLayout from "../layouts/NormalLayout"
-import H2 from "../typography/H2"
+import CompactLayout from "../../components/layouts/CompactLayout"
+import H2 from "../../components/typography/H2"
+import { Button } from "../../components/ui/button"
 import {
     Form,
     FormControl,
@@ -8,19 +9,14 @@ import {
     FormItem,
     FormLabel,
     FormMessage,
-} from "../ui/form"
+} from "../../components/ui/form"
+import { Input } from "../../components/ui/input"
 import { useForm } from "react-hook-form"
 import { zodResolver } from "@hookform/resolvers/zod"
-import { updateQuickLink } from "./QuickLinkManager"
-import CompactLayout from "../layouts/CompactLayout"
-import { Input } from "../ui/input"
-import { Button } from "../ui/button"
-import { QuickLink } from "@src/types/quickLinkTypes"
-
-interface EditQuickLinkFormProps {
-    quickLink: QuickLink
-    afterSubmit?: () => void
-}
+import NormalLayout from "../../components/layouts/NormalLayout"
+import { addQuickLink } from "./QuickLinkManager"
+import { getModuleEmoji } from "@src/features/contentEnhancers/emoji/modules"
+import { extractUrlSegments } from "@src/features/versionControl"
 
 const formSchema = z.object({
     icon: z.string().emoji().max(4, "4 emoji max"),
@@ -28,19 +24,32 @@ const formSchema = z.object({
     href: z.string().min(1, "Link is required"),
 })
 
-export default function EditQuickLinkForm({
-    quickLink,
+interface AddQuickLinkFormProps {
+    name?: string
+    href?: string
+    afterSubmit?: () => void
+}
+
+export default function AddQuickLinkForm({
+    href,
+    name,
     afterSubmit,
-}: EditQuickLinkFormProps) {
+}: AddQuickLinkFormProps) {
+    const urlSegments = extractUrlSegments(href || "")
+    const moduleCode = urlSegments[0]
+    const moduleEmoji = getModuleEmoji(moduleCode)
+
     const form = useForm<z.infer<typeof formSchema>>({
         resolver: zodResolver(formSchema),
         defaultValues: {
-            ...quickLink,
+            icon: moduleEmoji,
+            name,
+            href,
         },
     })
 
     async function onSubmit(quickLinkData: z.infer<typeof formSchema>) {
-        await updateQuickLink(quickLink.id, quickLinkData)
+        await addQuickLink(quickLinkData)
 
         if (afterSubmit) {
             afterSubmit()
@@ -49,7 +58,7 @@ export default function EditQuickLinkForm({
 
     return (
         <NormalLayout>
-            <H2>Update &quot;{quickLink.name}&quot;</H2>
+            <H2>Add quick link</H2>
             <Form {...form}>
                 <form onSubmit={form.handleSubmit(onSubmit)} className="m-0">
                     <NormalLayout>
@@ -115,9 +124,7 @@ export default function EditQuickLinkForm({
                                 }}
                             />
                         </CompactLayout>
-                        <div className="flex flex-col-reverse sm:flex-row sm:justify-end sm:space-x-2">
-                            <Button type="submit">Save</Button>
-                        </div>
+                        <Button type="submit">Add</Button>
                     </NormalLayout>
                 </form>
             </Form>
