+++ src/features/quickAccess/QuickLinkContainer.tsx
@@ -2,8 +2,6 @@ interface QuickLinkContainerProps {
     children: React.ReactNode
 }
 
-export default function QuickLinkContainer({
-    children,
-}: QuickLinkContainerProps) {
+export function QuickLinkContainer({ children }: QuickLinkContainerProps) {
     return <div className="flex flex-wrap gap-3">{children}</div>
 }
