+++ src/pages/content/pages/Folder.tsx
@@ -7,7 +7,7 @@ import SubheaderBreadcrumbs from "@src/components/header/SubheaderBreadCrumbs"
 import CompactLayout from "@src/components/layouts/CompactLayout"
 import MainLayout from "@src/components/layouts/MainLayout"
 import WideLayout from "@src/components/layouts/WideLayout"
-import QuickLinks from "@src/components/quickAccess/QuickLinks"
+import { QuickLinks } from "@src/features/quickAccess/"
 import Table from "@src/components/table/Table"
 import { ThemeToggle } from "@src/components/ui/theme-toggle"
 import { FileMetricsTracker } from "@src/features/versionControl"
